添加生成库存的接口

+++ eis-jianming-facade/src/main/java/com/prolog/eis/model/wms/ContainerTask.java
@@ -206,7 +206,7 @@ public class ContainerTask {
 		this.ownerId = ownerId;
 	}
 
-	public double getQty() {
+	public float getQty() {
 		return qty;
 	}
